+++ scelvis/ui/common.py
@@ -28,14 +28,15 @@ def render_plot(data_type, plot_type):
                     target="_blank",
                 ),
             ],
-            plot_type == "scatter",
+            plot_type != "scatter",
         )
 
 
 def render_subsampling_dropdown(data, token):
-    sample_choices = [{"label": g, "value": g} for g in [1000, 5000] if g < data.meta.shape[0]] + [
-        {"label": "all", "value": "all"}
-    ]
+
+    sample_choices = [
+        {"label": g, "value": g} for g in [1000, 5000] if g < data.ad.obs.shape[0]
+    ] + [{"label": "all", "value": "all"}]
     return html.Div(
         [
             html.Label("select cell sample"),