+++ scelvis/ui/genes.py
@@ -9,7 +9,7 @@ import dash_table
 import numpy as np
 import pandas as pd
 import plotly.graph_objs as go
-import plotly.tools as tools
+import plotly.subplots as subplots
 
 from .. import settings
 from . import colors, common
@@ -24,13 +24,13 @@ def render_controls_scatter(data):
                 dcc.Dropdown(
                     id="expression_scatter_select_x",
                     options=[{"label": c, "value": c} for c in data.numerical_meta],
-                    value=data.meta.columns[0],
+                    value=data.numerical_meta[0],
                 ),
                 html.Label("select y axis"),
                 dcc.Dropdown(
                     id="expression_scatter_select_y",
                     options=[{"label": c, "value": c} for c in data.numerical_meta],
-                    value=data.meta.columns[1],
+                    value=data.numerical_meta[1],
                 ),
             ],
             title="""select x- and y-coordinates of embedding (e.g., TSNE or UMAP)""",
@@ -78,7 +78,7 @@ def render_controls_dot(data):
                 options=[
                     {"label": c, "value": c}
                     for c in data.categorical_meta
-                    if len(data.meta[c].unique()) < 4
+                    if len(data.ad.obs[c].unique()) < 4
                 ],
                 value=None,
             ),
@@ -138,7 +138,7 @@ def render_controls(data):
                             "disabled": (data.markers is None),
                         }
                     ],
-                    values=[],
+                    value=[],
                     className="mt-2",
                     inputClassName="mr-1",
                 ),
@@ -178,7 +178,7 @@ def render(data):
 
 
 def render_marker_list(data, values):
-    if "markers" in values:
+    if values is not None and "markers" in values:
         return [
             dbc.Row([dbc.Col(html.Hr())]),
             dbc.Row(
@@ -193,12 +193,11 @@ def render_marker_list(data, values):
                                 id="marker_list",
                                 columns=[{"name": i, "id": i} for i in data.markers.columns],
                                 data=data.markers.round(3).to_dict("rows"),
-                                n_fixed_rows=1,
+                                fixed_rows={"headers": True},
                                 style_as_list_view=True,
-                                sorting=True,
-                                sorting_type="single",
+                                sort_action="native",
+                                sort_mode="single",
                                 row_selectable="multi",
-                                pagination_mode=False,
                                 selected_rows=[],
                                 style_table={"maxHeight": 200, "overflowY": "scroll"},
                             )
@@ -210,22 +209,20 @@ def render_marker_list(data, values):
         ]
 
     else:
-        return [html.P()]
+        return [dbc.Row([dbc.Col(html.P())]), html.P()]
 
 
 def render_plot_scatter(data, xc, yc, genelist, sample_size):
 
-    gl = [g for g in genelist if g in data.DGE.index]
+    gl = [g for g in genelist if g in data.ad.var_names]
 
     if gl is None or len(gl) == 0 or xc is None or yc is None:
         return {}, "", True
 
     if sample_size != "all":
-        DGE_here = data.DGE.sample(n=sample_size, axis=1)
-        meta_here = data.meta.loc[DGE_here.columns]
+        ad_here = data.ad[np.random.choice(data.ad.obs_names, sample_size, replace=False), :]
     else:
-        DGE_here = data.DGE
-        meta_here = data.meta
+        ad_here = data.ad
 
     ngenes = len(gl)
     if ngenes > 1:
@@ -235,30 +232,30 @@ def render_plot_scatter(data, xc, yc, genelist, sample_size):
         nrow = 1
         ncol = 1
 
-    fig = tools.make_subplots(
+    fig = subplots.make_subplots(
         rows=nrow,
         cols=ncol,
         specs=[[{} for c in range(ncol)] for r in range(nrow)],
         shared_xaxes=True,
         shared_yaxes=True,
-        vertical_spacing=0.01,
-        horizontal_spacing=0.01,
+        vertical_spacing=0.05,
+        horizontal_spacing=0.05,
         subplot_titles=gl,
     )
 
     # rescale all expression values to the same min and max
-    maxval = max(1, DGE_here.loc[gl].max().max())
+    maxval = max(1, ad_here[:, gl].X.max().max())
 
     for ng, gene in enumerate(gl):
         # for numerical data, plot scatter all at once
         trace = go.Scatter(
-            x=meta_here[xc],
-            y=meta_here[yc],
-            text=data.cells,
+            x=ad_here.obs[xc],
+            y=ad_here.obs[yc],
+            text=ad_here.obs_names,
             mode="markers",
             marker={
                 "size": 3,
-                "color": maxval * DGE_here.loc[gene] / max(1, DGE_here.loc[gene].max()),
+                "color": maxval * ad_here[:, gene].X / max(1, ad_here[:, gene].X.max()),
                 "colorscale": "Viridis",
                 "colorbar": {"title": "expression", "titleside": "right"},
                 "showscale": (ng == 0),
@@ -268,22 +265,27 @@ def render_plot_scatter(data, xc, yc, genelist, sample_size):
         nc = ng % ncol + 1
         nr = ng // ncol + 1
         fig.append_trace(trace, nr, nc)
+        if nc == 1:
+            fig["layout"]["yaxis" + (str(ng + 1) if ng > 0 else "")].update(title=yc)
+        if nr == nrow:
+            fig["layout"]["xaxis" + (str(ng + 1) if ng > 0 else "")].update(title=xc)
 
     # fix axes labels and scales
-    for k in dir(fig["layout"]):
-        if k.startswith("yaxis"):
-            fig["layout"][k].update(title=yc)
-        elif k.startswith("xaxis"):
-            fig["layout"][k].update(title=xc)
+    # for k in dir(fig["layout"]):
+    #    if k.startswith("yaxis"):
+    #        fig["layout"][k].update(title=yc)
+    #    elif k.startswith("xaxis"):
+    #        fig["layout"][k].update(title=xc)
 
     fig["layout"].update(
         margin={"l": 40, "b": 40, "t": 40, "r": 40},
+        plot_bgcolor="rgb(255,255,255)",
         showlegend=False,
         hovermode="closest",
         height=settings.PLOT_HEIGHT,
     )
 
-    plot_data = data.DGE.loc[gl].T.to_csv(index=True, header=True, encoding="utf-8")
+    plot_data = data.ad[:, gl].to_df().to_csv(index=True, header=True, encoding="utf-8")
     csv_string = "data:text/csv;charset=utf-8," + urllib.parse.quote(plot_data)
 
     return fig, csv_string, False
@@ -291,29 +293,27 @@ def render_plot_scatter(data, xc, yc, genelist, sample_size):
 
 def render_plot_violin(data, pathname, genelist, sample_size, group, split):
 
-    gl = [g for g in genelist if g in data.DGE.index]
+    gl = [g for g in genelist if g in data.ad.var_names]
 
     if gl is None or len(gl) == 0 or group is None:
         return {}, "", True
 
     if sample_size != "all":
-        DGE_here = data.DGE.sample(n=sample_size, axis=1)
-        meta_here = data.meta.loc[DGE_here.columns]
+        ad_here = data.ad[np.random.choice(data.ad.obs_names, sample_size, replace=False), :]
     else:
-        DGE_here = data.DGE
-        meta_here = data.meta
+        ad_here = data.ad
 
     # select color palette
     if split is None:
-        groupvals = meta_here[group].cat.categories
+        groupvals = ad_here.obs[group].cat.categories
         cm = colors.get_cm(groupvals)
     else:
-        splitvals = meta_here[split].cat.categories
+        splitvals = ad_here.obs[split].cat.categories
         cm = colors.get_cm(splitvals)
 
     ngenes = len(gl)
 
-    fig = tools.make_subplots(
+    fig = subplots.make_subplots(
         rows=ngenes, cols=1, specs=[[{}] for gene in gl], shared_xaxes=True, vertical_spacing=0.01
     )
     sg = 0
@@ -321,7 +321,7 @@ def render_plot_violin(data, pathname, genelist, sample_size, group, split):
 
         if split is None:
             for n, cv in enumerate(groupvals):
-                y = DGE_here.loc[gene, meta_here[group] == cv]
+                y = ad_here[ad_here.obs[group] == cv, :][:, gene].X
                 tr = go.Violin(
                     y=y,
                     name=cv,
@@ -335,12 +335,12 @@ def render_plot_violin(data, pathname, genelist, sample_size, group, split):
                 )
                 fig.append_trace(tr, ngenes - ng, 1)
                 sg += 1
-            plot_data = data.DGE.loc[gl].T.join(data.meta[group])
+            plot_data = data.ad[:, gl].to_df().join(data.ad.obs[group])
         else:
             for n, sv in enumerate(splitvals):
-                y = DGE_here.loc[gene, meta_here[split] == sv]
+                y = ad_here[ad_here.obs[split] == sv, :][:, gene].X
                 tr = go.Violin(
-                    x=meta_here[meta_here[split] == sv][group],
+                    x=ad_here[ad_here.obs[split] == sv].obs[group],
                     y=y,
                     name=sv,
                     offsetgroup=n,
@@ -354,7 +354,7 @@ def render_plot_violin(data, pathname, genelist, sample_size, group, split):
                 )
                 fig.append_trace(tr, ngenes - ng, 1)
                 sg += 1
-            plot_data = data.DGE.loc[gl].T.join(data.meta[[group, split]])
+            plot_data = data.ad[:, gl].to_df().join(data.ad.obs[[group, split]])
 
     for ng, gene in enumerate(gl):
         if ngenes - ng == 1:
@@ -362,8 +362,11 @@ def render_plot_violin(data, pathname, genelist, sample_size, group, split):
         else:
             fig["layout"]["yaxis" + str(ngenes - ng)].update(title=gene)
 
+    fig["layout"]["xaxis" + (str(ngenes) if ngenes > 1 else "")].update(title=group)
+
     fig["layout"].update(
-        xaxis={"title": group, "tickangle": -45},
+        xaxis={"tickangle": -45},
+        plot_bgcolor="rgb(255,255,255)",
         margin={"l": 50, "b": 80, "t": 10, "r": 10},
         legend={"x": 1.05, "y": 1},
         hovermode="closest",
@@ -382,12 +385,12 @@ def render_plot_violin(data, pathname, genelist, sample_size, group, split):
 
 def render_plot_dot(data, pathname, genelist, group, split):
 
-    gl = [g for g in genelist if g in data.DGE.index]
+    gl = [g for g in genelist if g in data.ad.var_names]
 
     if gl is None or len(gl) == 0 or group is None:
         return {}, "", True
 
-    groupvals = data.meta[group].cat.categories
+    groupvals = data.ad.obs[group].cat.categories
     ngroup = len(groupvals)
     ngenes = len(gl)
 
@@ -396,7 +399,12 @@ def render_plot_dot(data, pathname, genelist, group, split):
 
     if split is None:
         plot_data = (
-            data.DGE.loc[gl].T.join(data.meta[group]).groupby(group).apply(my_agg).unstack(level=1)
+            data.ad[:, gl]
+            .to_df()
+            .join(data.ad.obs[group])
+            .groupby(group)
+            .apply(my_agg)
+            .unstack(level=1)
         )
         means = plot_data.xs("expression", axis=1, level=1).values
         sizes = plot_data.xs("pct_cells", axis=1, level=1).values
@@ -418,12 +426,11 @@ def render_plot_dot(data, pathname, genelist, group, split):
         # extra invisible traces for legend
         traces += [
             go.Scatter(
-                x=[0],
-                y=[0],
+                x=[None],
+                y=[None],
                 mode="markers",
                 marker={"size": 20 * s, "color": "black"},
                 name=p,
-                visible="legendonly",
                 legendgroup="size",
                 showlegend=True,
             )
@@ -431,12 +438,11 @@ def render_plot_dot(data, pathname, genelist, group, split):
         ]
         traces += [
             go.Scatter(
-                x=[0],
-                y=[0],
+                x=[None],
+                y=[None],
                 mode="markers",
                 marker={"size": 20, "color": colors.my_gradients[0][c][1], "symbol": "square"},
                 name=p,
-                visible="legendonly",
                 legendgroup="color",
                 showlegend=True,
             )
@@ -465,15 +471,22 @@ def render_plot_dot(data, pathname, genelist, group, split):
         )
 
     else:
-        splitvals = data.meta[split].cat.categories
+        splitvals = data.ad.obs[split].cat.categories
         nsplit = len(splitvals)
         traces = []
+        # fix bug in pandas groupby when having more than one grouping variable
+        ii = pd.MultiIndex.from_arrays(
+            pd.core.reshape.util.cartesian_product([groupvals, splitvals])
+        )
         plot_data = (
-            data.DGE.loc[gl]
-            .T.join(data.meta[[group, split]])
+            data.ad[:, gl]
+            .to_df()
+            .join(data.ad.obs[[group, split]])
             .groupby([group, split])
             .apply(my_agg)
             .unstack(level=2)
+            .reindex(ii)
+            .fillna(0)
         )
         for n, sv in enumerate(splitvals):
             means = plot_data.xs("expression", axis=1, level=1).xs(sv, axis=0, level=1).values
@@ -501,7 +514,6 @@ def render_plot_dot(data, pathname, genelist, group, split):
                 mode="markers",
                 marker={"size": 20 * s, "color": "black"},
                 name=p,
-                visible="legendonly",
                 legendgroup="size",
                 showlegend=True,
             )
@@ -514,7 +526,6 @@ def render_plot_dot(data, pathname, genelist, group, split):
                 mode="markers",
                 marker={"size": 20, "symbol": "square", "color": colors.my_gradients[-1][c][1]},
                 name=p,
-                visible="legendonly",
                 legendgroup="color",
                 showlegend=True,
             )
@@ -527,7 +538,6 @@ def render_plot_dot(data, pathname, genelist, group, split):
                 mode="markers",
                 marker={"size": 20, "symbol": "square", "color": colors.my_gradients[n][1][1]},
                 name=sv,
-                visible="legendonly",
                 legendgroup="split",
                 showlegend=True,
             )
@@ -554,6 +564,7 @@ def render_plot_dot(data, pathname, genelist, group, split):
             showlegend=True,
             hovermode="closest",
             height=settings.PLOT_HEIGHT,
+            plot_bgcolor="rgb(255,255,255)",
         )
 
     csv_string = "data:text/csv;charset=utf-8," + urllib.parse.quote(